world3d: Separated voxelmap data from approach impl. -> less redundancy.
