+++ plugins/clay_world/voxelmapgeometry.cpp
@@ -20,259 +20,98 @@ static QColor stringToColor(const QString &str)
 
 VoxelMapGeometry::VoxelMapGeometry()
 {
+    // Connect the data change notification to trigger geometry updates
+    m_data.setOnDataChanged([this]() { updateGeometry(); });
+
+    // Connect property change signals
+    connect(&m_data, &VoxelMapData::widthChanged, this, &VoxelMapGeometry::widthChanged);
+    connect(&m_data, &VoxelMapData::heightChanged, this, &VoxelMapGeometry::heightChanged);
+    connect(&m_data, &VoxelMapData::depthChanged, this, &VoxelMapGeometry::depthChanged);
+    connect(&m_data, &VoxelMapData::voxelSizeChanged, this, &VoxelMapGeometry::voxelSizeChanged);
+    connect(&m_data, &VoxelMapData::spacingChanged, this, &VoxelMapGeometry::spacingChanged);
+}
+
+int VoxelMapGeometry::width() const
+{
+    return m_data.width();
 }
 
 void VoxelMapGeometry::setWidth(int w)
 {
-    if (m_width == w)
-        return;
-    m_width = w;
-    m_voxels.fill(Qt::transparent, m_width * m_height * m_depth);
-    emit widthChanged();
-    updateGeometry();
+    m_data.setWidth(w);
+}
+
+int VoxelMapGeometry::height() const
+{
+    return m_data.height();
 }
 
 void VoxelMapGeometry::setHeight(int h)
 {
-    if (m_height == h)
-        return;
-    m_height = h;
-    m_voxels.fill(Qt::transparent, m_width * m_height * m_depth);
-    emit heightChanged();
-    updateGeometry();
+    m_data.setHeight(h);
+}
+
+int VoxelMapGeometry::depth() const
+{
+    return m_data.depth();
 }
 
 void VoxelMapGeometry::setDepth(int d)
 {
-    if (m_depth == d)
-        return;
-    m_depth = d;
-    m_voxels.fill(Qt::transparent, m_width * m_height * m_depth);
-    emit depthChanged();
-    updateGeometry();
+    m_data.setDepth(d);
+}
+
+float VoxelMapGeometry::voxelSize() const
+{
+    return m_data.voxelSize();
 }
 
 void VoxelMapGeometry::setVoxelSize(float size)
 {
-    if (qFuzzyCompare(m_voxelSize, size))
-        return;
-    m_voxelSize = size;
-    emit voxelSizeChanged();
-    updateGeometry();
+    m_data.setVoxelSize(size);
 }
 
-QColor VoxelMapGeometry::voxel(int x, int y, int z) const
+float VoxelMapGeometry::spacing() const
 {
-    if (x<0 || x>=m_width || y<0 || y>=m_height || z<0 || z>=m_depth)
-        return Qt::transparent;
-    return m_voxels[indexOf(x,y,z)];
+    return m_data.spacing();
 }
 
-void VoxelMapGeometry::setVoxel(int x, int y, int z, const QColor &color)
+void VoxelMapGeometry::setSpacing(float spacing)
 {
-    if (x<0 || x>=m_width || y<0 || y>=m_height || z<0 || z>=m_depth)
-        return;
-    int idx = indexOf(x,y,z);
-    if (m_voxels[idx] == color)
-        return;
-    m_voxels[idx] = color;
-    updateGeometry();
+    m_data.setSpacing(spacing);
 }
 
-namespace
+// ==========================================
+// Delegated Methods (for QML-invokable functions)
+// ==========================================
+bool VoxelMapGeometry::saveToFile(const QString &path, bool binary)
 {
-
-struct ColorProb {
-    QColor color;
-    float probability;
-};
-
-QVector<ColorProb> prepareColorDistribution(const QVariantList &colorDistribution) {
-    QVector<ColorProb> distribution;
-    float totalWeight = 0.0f;
-
-    // Process each item in the array
-    for (const QVariant &item : colorDistribution) {
-        QVariantMap entry = item.toMap();
-        if (entry.contains("color") && entry.contains("weight")) {
-            QColor color = QColor(entry["color"].toString());
-            float weight = entry["weight"].toFloat();
-            if (weight > 0.0f) {
-                totalWeight += weight;
-                distribution.append({color, weight});
-            }
-        }
-    }
-
-    if (!distribution.isEmpty()) {
-        // Normalize probabilities
-        for (auto &item : distribution) {
-            item.probability /= totalWeight;
-        }
-    }
-
-    return distribution;
+    return m_data.saveToFile(path, binary);
 }
 
-QColor getRandomColor(const QVector<ColorProb> &distribution) {
-    static std::random_device rd;
-    static std::mt19937 gen(rd());
-    static std::uniform_real_distribution<float> dis(0.0f, 1.0f);
-
-    float rand = dis(gen);
-    float cumulative = 0.0f;
-    QColor selectedColor = distribution[0].color; // fallback
-
-    for (const auto &item : distribution) {
-        cumulative += item.probability;
-        if (rand <= cumulative) {
-            selectedColor = item.color;
-            break;
-        }
-    }
-
-    return selectedColor;
-}
-
-float applyNoise(float value, float noiseFactor) {
-    if (noiseFactor <= 0.0f) return value;
-
-    static std::random_device rd;
-    static std::mt19937 gen(rd());
-    std::uniform_real_distribution<float> dis(-noiseFactor, noiseFactor);
-
-    return value * (1.0f + dis(gen));
-}
-
-}
-
-void VoxelMapGeometry::fillSphere(int cx, int cy, int cz, int r, const QVariantList &colorDistribution, float noiseFactor)
+bool VoxelMapGeometry::loadFromFile(const QString &path, bool binary)
 {
-    // Validate inputs
-    if (r <= 0.0f || colorDistribution.isEmpty()) return;
-    if (cx < -r || cy < -r || cz < -r) return;
-    if (cx >= width() + r || cy >= height() + r || cz >= depth() + r) return;
-
-    auto distribution = prepareColorDistribution(colorDistribution);
-    if (distribution.isEmpty()) return;
-
-    // Use maximum possible radius for bounds checking
-    float maxRadius = r * (1.0f + noiseFactor);
-    float baseR2 = (r - 0.5f) * (r - 0.5f);
-
-    int minX = qBound(0, int(cx - maxRadius), width() - 1);
-    int maxX = qBound(0, int(cx + maxRadius), width() - 1);
-    int minY = qBound(0, int(cy - maxRadius), height() - 1);
-    int maxY = qBound(0, int(cy + maxRadius), height() - 1);
-    int minZ = qBound(0, int(cz - maxRadius), depth() - 1);
-    int maxZ = qBound(0, int(cz + maxRadius), depth() - 1);
-
-    if (minX > maxX || minY > maxY || minZ > maxZ) return;
-
-    for (int z = minZ; z <= maxZ; ++z) {
-        for (int y = minY; y <= maxY; ++y) {
-            for (int x = minX; x <= maxX; ++x) {
-                float dx = float(x - cx);
-                float dy = float(y - cy);
-                float dz = float(z - cz);
-                float distanceSquared = dx*dx + dy*dy + dz*dz;
-                float currentR2 = applyNoise(baseR2, noiseFactor);
-                if (distanceSquared <= currentR2) {
-                    m_voxels[indexOf(x,y,z)] = getRandomColor(distribution);
-                }
-            }
-        }
-    }
-
-    updateGeometry();
+    return m_data.loadFromFile(path, binary);
 }
 
-void VoxelMapGeometry::fillCylinder(int cx, int cy, int cz, int r, int height, const QVariantList &colorDistribution, float noiseFactor)
-{
-    // Validate inputs
-    if (r <= 0.0f || height <= 0 || colorDistribution.isEmpty()) return;
-    if (cx < -r || cy < 0 || cz < -r) return;
-    if (cx >= width() + r || cy >= this->height() + height || cz >= depth() + r) return;
-
-    auto distribution = prepareColorDistribution(colorDistribution);
-    if (distribution.isEmpty()) return;
-
-    // Use maximum possible radius for bounds checking
-    float maxRadius = r * (1.0f + noiseFactor);
-    float baseR2 = (r - 0.5f) * (r - 0.5f);
-
-    int minX = qBound(0, int(cx - maxRadius), width() - 1);
-    int maxX = qBound(0, int(cx + maxRadius), width() - 1);
-    int minY = qBound(0, int(cy), this->height() - 1);
-    int maxY = qBound(0, int(cy + height), this->height() - 1);
-    int minZ = qBound(0, int(cz - maxRadius), depth() - 1);
-    int maxZ = qBound(0, int(cz + maxRadius), depth() - 1);
-
-    if (minX > maxX || minY > maxY || minZ > maxZ) return;
-
-    for (int y = minY; y <= maxY; ++y) {
-        float currentR2 = applyNoise(baseR2, noiseFactor);
-        for (int z = minZ; z <= maxZ; ++z) {
-            for (int x = minX; x <= maxX; ++x) {
-                float dx = float(x - cx);
-                float dz = float(z - cz);
-                if (dx*dx + dz*dz <= currentR2) {
-                    m_voxels[indexOf(x,y,z)] = getRandomColor(distribution);
-                }
-            }
-        }
-    }
-
-    updateGeometry();
+QColor VoxelMapGeometry::voxel(int x, int y, int z) const {
+    return m_data.voxel(x, y, z);
 }
 
-void VoxelMapGeometry::fillBox(int cx, int cy, int cz, int boxWidth, int boxHeight, int boxDepth, const QVariantList &colorDistribution, float noiseFactor)
-{
-    // Validate inputs
-    if (boxWidth <= 0 || boxHeight <= 0 || boxDepth <= 0 || colorDistribution.isEmpty()) return;
+void VoxelMapGeometry::setVoxel(int x, int y, int z, const QColor &color) {
+    m_data.setVoxel(x, y, z, color);
+}
 
-    auto distribution = prepareColorDistribution(colorDistribution);
-    if (distribution.isEmpty()) return;
+void VoxelMapGeometry::fillSphere(int cx, int cy, int cz, int r, const QVariantList &colorDistribution, float noiseFactor) {
+    m_data.fillSphere(cx, cy, cz, r, colorDistribution, noiseFactor);
+}
 
-    // Calculate box bounds with noise factor
-    float halfWidth = boxWidth / 2.0f;
-    float halfHeight = boxHeight / 2.0f;
-    float halfDepth = boxDepth / 2.0f;
+void VoxelMapGeometry::fillCylinder(int cx, int cy, int cz, int r, int height, const QVariantList &colorDistribution, float noiseFactor) {
+    m_data.fillCylinder(cx, cy, cz, r, height, colorDistribution, noiseFactor);
+}
 
-    // Calculate boundaries in voxel coordinates
-    int minX = qBound(0, int(cx - halfWidth), width() - 1);
-    int maxX = qBound(0, int(cx + halfWidth), width() - 1);
-    int minY = qBound(0, int(cy - halfHeight), height() - 1);
-    int maxY = qBound(0, int(cy + halfHeight), height() - 1);
-    int minZ = qBound(0, int(cz - halfDepth), depth() - 1);
-    int maxZ = qBound(0, int(cz + halfDepth), depth() - 1);
-
-    if (minX > maxX || minY > maxY || minZ > maxZ) return;
-
-    // Fill the box
-    for (int z = minZ; z <= maxZ; ++z) {
-        for (int y = minY; y <= maxY; ++y) {
-            for (int x = minX; x <= maxX; ++x) {
-                // Apply noise to the box boundaries
-                if (noiseFactor > 0.0f) {
-                    float dx = float(x - cx) / halfWidth;
-                    float dy = float(y - cy) / halfHeight;
-                    float dz = float(z - cz) / halfDepth;
-
-                    // Skip if outside the noisy boundary
-                    if (std::abs(dx) > (1.0f + applyNoise(0.0f, noiseFactor)) ||
-                        std::abs(dy) > (1.0f + applyNoise(0.0f, noiseFactor)) ||
-                        std::abs(dz) > (1.0f + applyNoise(0.0f, noiseFactor))) {
-                        continue;
-                    }
-                }
-
-                m_voxels[indexOf(x,y,z)] = getRandomColor(distribution);
-            }
-        }
-    }
-
-    updateGeometry();
+void VoxelMapGeometry::fillBox(int cx, int cy, int cz, int width, int height, int depth, const QVariantList &colorDistribution, float noiseFactor) {
+    m_data.fillBox(cx, cy, cz, width, height, depth, colorDistribution, noiseFactor);
 }
 
 bool VoxelMapGeometry::isFaceVisible(int x, int y, int z, int faceIndex) const {
@@ -288,33 +127,24 @@ bool VoxelMapGeometry::isFaceVisible(int x, int y, int z, int faceIndex) const {
     }
 
     // If neighbor is outside bounds, face is visible
-    if (nx < 0 || nx >= m_width || ny < 0 || ny >= m_height || nz < 0 || nz >= m_depth)
+    if (nx < 0 || nx >= m_data.width() || ny < 0 || ny >= m_data.height() || nz < 0 || nz >= m_data.depth())
         return true;
 
     // Face is visible if neighbor voxel is transparent
-    return m_voxels[indexOf(nx, ny, nz)].alpha() == 0;
-}
-
-void VoxelMapGeometry::setSpacing(float spacing)
-{
-    if (qFuzzyCompare(m_spacing, spacing))
-        return;
-    m_spacing = spacing;
-    emit spacingChanged();
-    updateGeometry();
+    return m_data.voxel(nx, ny, nz).alpha() == 0;
 }
 
 // Build all voxel cubes in a single geometry
 void VoxelMapGeometry::updateGeometry()
 {
     clear();
-    if (m_width <= 0 || m_height <= 0 || m_depth <= 0)
+    if (m_data.width() <= 0 || m_data.height() <= 0 || m_data.depth() <= 0)
         return;
 
     // Calculate the total size including spacing
-    float totalWidth = m_width * (m_voxelSize + m_spacing) - m_spacing;
-    float totalHeight = m_height * (m_voxelSize + m_spacing) - m_spacing;
-    float totalDepth = m_depth * (m_voxelSize + m_spacing) - m_spacing;
+    float totalWidth = m_data.width() * (m_data.voxelSize() + m_data.spacing()) - m_data.spacing();
+    float totalHeight = m_data.height() * (m_data.voxelSize() + m_data.spacing()) - m_data.spacing();
+    float totalDepth = m_data.depth() * (m_data.voxelSize() + m_data.spacing()) - m_data.spacing();
 
     // Update bounds calculation
     float halfWidth = totalWidth / 2.0f;
@@ -326,9 +156,9 @@ void VoxelMapGeometry::updateGeometry()
     QByteArray vertexBuffer;
     QByteArray indexBuffer;
     // 24 vertices per cube (4 vertices per face * 6 faces)
-    vertexBuffer.reserve(m_width * m_height * m_depth * 24 * sizeof(float) * 10);
+    vertexBuffer.reserve(m_data.width() * m_data.height() * m_data.depth() * 24 * sizeof(float) * 10);
     // 36 indices per cube (6 indices per face * 6 faces)
-    indexBuffer.reserve(m_width * m_height * m_depth * 36 * sizeof(quint32));
+    indexBuffer.reserve(m_data.width() * m_data.height() * m_data.depth() * 36 * sizeof(quint32));
 
     int vertexCount = 0;
 
@@ -346,17 +176,17 @@ void VoxelMapGeometry::updateGeometry()
     float offsetX = -totalWidth / 2.0f;
     float offsetZ = -totalDepth / 2.0f;
 
-    for (int z = 0; z < m_depth; ++z) {
-        for (int y = 0; y < m_height; ++y) {
-            for (int x = 0; x < m_width; ++x) {
-                QColor c = m_voxels[indexOf(x,y,z)];
+    for (int z = 0; z < m_data.depth(); ++z) {
+        for (int y = 0; y < m_data.height(); ++y) {
+            for (int x = 0; x < m_data.width(); ++x) {
+                QColor c = m_data.voxel(x,y,z);
                 if (c.alpha() == 0) continue;
 
                 // Update position calculations to include spacing
-                float fx = offsetX + x * (m_voxelSize + m_spacing);
-                float fy = y * (m_voxelSize + m_spacing);
-                float fz = offsetZ + z * (m_voxelSize + m_spacing);
-                float s = m_voxelSize;
+                float fx = offsetX + x * (m_data.voxelSize() + m_data.spacing());
+                float fy = y * (m_data.voxelSize() + m_data.spacing());
+                float fz = offsetZ + z * (m_data.voxelSize() + m_data.spacing());
+                float s = m_data.voxelSize();
 
                 // Define vertices for each face (4 vertices per face)
                 struct Face {
@@ -430,100 +260,4 @@ void VoxelMapGeometry::updateGeometry()
 
     setPrimitiveType(QQuick3DGeometry::PrimitiveType::Triangles);
     update(); // finalize
-}
-
-bool VoxelMapGeometry::saveToFile(const QString &path, bool binary)
-{
-    QFile file(path);
-    if (!file.open(QIODevice::WriteOnly)) {
-        qWarning() << "Could not open for writing:" << path;
-        return false;
-    }
-    if (binary) {
-        QDataStream out(&file);
-        out << m_width << m_height << m_depth << m_voxelSize;
-        for (const auto &col : m_voxels) {
-            out << col.rgba(); // store 32-bit RGBA
-        }
-    } else {
-        QTextStream ts(&file);
-        ts << "width: " << m_width << "\n"
-           << "height: " << m_height << "\n"
-           << "depth: " << m_depth << "\n"
-           << "voxelSize: " << m_voxelSize << "\n"
-           << "voxels:\n";
-        for (int z=0; z<m_depth; ++z) {
-            for (int y=0; y<m_height; ++y) {
-                for (int x=0; x<m_width; ++x) {
-                    QColor c = m_voxels[indexOf(x,y,z)];
-                    if (c.alpha() != 0) {
-                        ts << "  - xyz: [" << x << "," << y << "," << z << "] color: "
-                           << colorToString(c) << "\n";
-                    }
-                }
-            }
-        }
-    }
-    return true;
-}
-
-bool VoxelMapGeometry::loadFromFile(const QString &path, bool binary)
-{
-    QFile file(path);
-    if (!file.open(QIODevice::ReadOnly)) {
-        qWarning() << "Could not open for reading:" << path;
-        return false;
-    }
-    if (binary) {
-        QDataStream in(&file);
-        in >> m_width >> m_height >> m_depth >> m_voxelSize;
-        m_voxels.resize(m_width * m_height * m_depth);
-        for (int i = 0; i < m_voxels.size(); ++i) {
-            QRgb val;
-            in >> val;
-            m_voxels[i].setRgba(val);
-        }
-    } else {
-        // Simple line-based parse for demonstration
-        QTextStream ts(&file);
-        m_voxels.clear();
-        m_width = m_height = m_depth = 0;
-        while (!ts.atEnd()) {
-            QString line = ts.readLine().trimmed();
-            if (line.startsWith("width:"))
-                m_width = line.split(":").last().trimmed().toInt();
-            else if (line.startsWith("height:"))
-                m_height = line.split(":").last().trimmed().toInt();
-            else if (line.startsWith("depth:"))
-                m_depth = line.split(":").last().trimmed().toInt();
-            else if (line.startsWith("voxelSize:"))
-                m_voxelSize = line.split(":").last().trimmed().toFloat();
-        }
-        // Rewind to parse voxel lines
-        file.seek(0);
-        ts.seek(0);
-        m_voxels.resize(m_width * m_height * m_depth);
-        while (!ts.atEnd()) {
-            QString line = ts.readLine().trimmed();
-            if (line.startsWith("- xyz:")) {
-                // example: "- xyz: [x,y,z] color: R,G,B,A"
-                int start = line.indexOf('[') + 1;
-                int end = line.indexOf(']');
-                QStringList coords = line.mid(start, end - start).split(',');
-                int x = coords[0].toInt();
-                int y = coords[1].toInt();
-                int z = coords[2].toInt();
-
-                int colorPos = line.indexOf("color:") + 6;
-                QString colorStr = line.mid(colorPos).trimmed();
-                setVoxel(x, y, z, stringToColor(colorStr));
-            }
-        }
-    }
-    emit widthChanged();
-    emit heightChanged();
-    emit depthChanged();
-    emit voxelSizeChanged();
-    updateGeometry();
-    return true;
 }
